Separated search panel helpers
