+++ cypress/integration/searchPage/results_filtering_spec.js
@@ -1,4 +1,4 @@
-import { SearchPanelFactory } from "../../support/SearchPanel";
+import { SearchPanelFactory } from "../../SearchPanel/SearchPanel";
 
 describe('Tests result filters', () => {
   let search;
