+++ cypress/integration/searchPage/search_params_in_url_spec.js
@@ -1,4 +1,4 @@
-import { SearchPanelFactory } from '../../support/SearchPanel';
+import { SearchPanelFactory } from '../../SearchPanel/SearchPanel';
 
 function getUrlWithQuery(keyword, filterLabel = undefined) {
   const encodedKeyword = encodeURI(keyword);
